+++ test/unit/timestamp.js
@@ -5,6 +5,14 @@ var sinon    = require('sinon');
 var Timestamp = require('../../src/timestamp');
 
 describe('Timestamp', function () {
+  var clock;
+  beforeEach(function () {
+    clock = sinon.useFakeTimers();
+  });
+  afterEach(function () {
+    clock.restore();
+  });
+
   describe('fromDate', function () {
     it('should convert from date', function () {
       var date = new Date('2009-02-13T23:31:30.123456789Z');
@@ -18,6 +26,13 @@ describe('Timestamp', function () {
       var timestamp = Timestamp.fromMillis(1234567890123);
       expect(timestamp.seconds).to.equal(1234567890);
       expect(timestamp.nanoseconds).to.equal(123000000);
+    });
+  });
+  describe('now', function () {
+    it('should create a current Timestamp', function () {
+      var now = new Date();
+      var timestamp = Timestamp.now();
+      expect(timestamp.toDate().getTime()).to.equal(now.getTime());
     });
   });
   describe('#toDate', function () {